1. 修改表情加载，但是还是不完善. 应该放入 Html.fromHtml 处理，但是只加载 placeholder, 随后真正 setText 显示的时候开始加载; 2. 修复推书频道链接点击无响应的问题.

+++ app/src/main/java/org/cryse/lkong/ui/adapter/TimelineAdapter.java
@@ -110,7 +110,7 @@ public class TimelineAdapter extends RecyclerViewBaseAdapter<TimelineModel> {
                 }
 
                 SimpleImageGetter imageGetter = new SimpleImageGetter(getContext(), ConnectionUtils.IMAGE_DOWNLOAD_ALWAYS)
-                        .setEmoticonSize(UIUtils.getSpDimensionPixelSize(getContext(), R.dimen.text_size_body1))
+                        .setEmoticonSize((int)UIUtils.getSpDimensionPixelSize(getContext(), R.dimen.text_size_body1))
                         .setPlaceHolder(R.drawable.image_placeholder)
                         .setError(R.drawable.image_placeholder);
                 Spanned spannedText = HtmlTextUtils.htmlToSpanned(mainContent, imageGetter, new HtmlTagHandler());
